LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/drawing-rectangles-to-the-canvas/2d.strokeRect.nonfinite.worker-expected.txt


PASS strokeRect() with Infinity/NaN is ignored
